+++ packages/reporter-svelte/src/index.js
@@ -1,3 +1,2 @@
 export * from './reporter';
-export * from './key';
 export { default as ValidationMessage } from './ValidationMessage.svelte';