+++ packages/site/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @felte/site
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @felte/reporter-svelte@0.3.0
+
 ## 0.0.4
 
 ### Patch Changes